Remove usages of RNSentry to a native wrapper (#857)
* ref: Move native crash to the native wrapper

Moved the native crash functionality to the native wrapper, along with
adding a utility helper function 'isModuleLoaded'. Includes tests for
the wrapper and the backend. Added __DEV__ global to jest to mock
react native functionality.

* ref: Moved boolean checks to wrapper

* ref: Completely migrated backend to use wrapper

Removed all instances of using the native module
RNSentry in ReactNativeBackend and now uses
the wrapper. Also includes the necessary tests and
an undefined dsn fallback. Tested on ios and android simulators.

* ref: Refactored fetchRelease to native wrapper

* ref: Refactored backend and release to promise resolve/reject

* ref: Moved deviceContexts to native wrapper

* test: Fixed startWithDsnString test with nativeClientAvailable check

* test: Added tests for sendEvent on both ios and android platforms

* ref: Added native client checks to every call to RNSentry

Added a check to all methods of the native wrapper that call
the RNSentry native module. Also used async/await syntax.