feat: Added event.origin tags to determine where the event originated.
The event.origin tag is added to the RN SDK Scope on init.
On Android, we have a switch
in the native beforeSend that determines where the event originated to set
the tag.
On iOS, the 'ios' tag is added in the native beforeSend if the event
originated in the sentry-cocoa sdk

Addresses #881