test: Add basic end-to-end tests workflow (#945)
Adds end to end tests that run on every PR and every push to master. Currently it only tests that the events sent from captureMessage and captureException are present on the Sentry API. The unhandled exception test is currently left out for now as on iOS the app crashes instantly with no time for the event id to be fetched by the tests, will look for a solution for this as soon as possible.

We should look to adding more tests to test the properties of the events later on.