feat: Touch events now track components with `sentry-label` (#2068)
Support tracking touch events on components with the prop `sentry-label`. 

Note that this does not throw a typescript error as ["it is not a valid JS Identifier"](https://www.typescriptlang.org/docs/handbook/jsx.html#attribute-type-checking)

It will fall back to other properties, with the hierarchy `sentry-label` > `accessibilityLabel` > `displayName`.

Also, removed the tracking of `name` on components without a displayName, found that doing so just caused components to be logged twice.